pyproject.toml
[tool.poetry]
name = "fastapibackend"
version = "0.1.0"
description = ""
authors = ["mikheyev9 <mikheyev9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.114.2"
uvicorn = "^0.30.6"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.34"
alembic = "^1.13.2"
pydantic = {extras = ["email"], version = "^2.9.1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"
python-multipart = "^0.0.9"
bcrypt = "4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
